+++ src/main/java/ua/goit/timonov/enterprise/module_6_2/controllers/DishController.java
@@ -2,10 +2,9 @@ package ua.goit.timonov.enterprise.module_6_2.controllers;
 
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import ua.goit.timonov.enterprise.module_6_2.model.DishDAO;
+import ua.goit.timonov.enterprise.module_6_2.dao.DishDAO;
 import ua.goit.timonov.enterprise.module_6_2.model.Dish;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
