+++ src/main/java/ua/goit/timonov/enterprise/module_6_2/controllers/OrderController.java
@@ -3,7 +3,7 @@ package ua.goit.timonov.enterprise.module_6_2.controllers;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import ua.goit.timonov.enterprise.module_6_2.model.Order;
-import ua.goit.timonov.enterprise.module_6_2.model.OrderDAO;
+import ua.goit.timonov.enterprise.module_6_2.dao.OrderDAO;
 
 import java.util.List;
