+++ src/main/java/ua/goit/timonov/enterprise/module_6_2/controllers/StorageController.java
@@ -3,8 +3,8 @@ package ua.goit.timonov.enterprise.module_6_2.controllers;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import ua.goit.timonov.enterprise.module_6_2.model.Ingredient;
-import ua.goit.timonov.enterprise.module_6_2.model.StorageDAO;
-import ua.goit.timonov.enterprise.module_6_2.model.jdbc.JdbcStorageDAO;
+import ua.goit.timonov.enterprise.module_6_2.dao.StorageDAO;
+import ua.goit.timonov.enterprise.module_6_2.dao.jdbc.JdbcStorageDAO;
 
 import java.util.List;
