+++ src/main/java/ua/goit/timonov/enterprise/module_6_2/dao/MenuDAO.java
@@ -1,4 +1,7 @@
-package ua.goit.timonov.enterprise.module_6_2.model;
+package ua.goit.timonov.enterprise.module_6_2.dao;
+
+import ua.goit.timonov.enterprise.module_6_2.model.Dish;
+import ua.goit.timonov.enterprise.module_6_2.model.Menu;
 
 import java.util.List;
