+++ src/main/java/ua/goit/timonov/enterprise/module_6_2/dao/jdbc/JdbcCookedDishDAO.java
@@ -1,8 +1,9 @@
-package ua.goit.timonov.enterprise.module_6_2.model.jdbc;
+package ua.goit.timonov.enterprise.module_6_2.dao.jdbc;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import ua.goit.timonov.enterprise.module_6_2.dao.CookedDishDAO;
 import ua.goit.timonov.enterprise.module_6_2.model.*;
 
 import java.util.ArrayList;
