+++ src/main/java/ua/goit/timonov/enterprise/module_6_2/dao/jdbc/JdbcEmployeeDAO.java
@@ -1,12 +1,10 @@
-package ua.goit.timonov.enterprise.module_6_2.model.jdbc;
+package ua.goit.timonov.enterprise.module_6_2.dao.jdbc;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import ua.goit.timonov.enterprise.module_6_2.model.Employee;
-import ua.goit.timonov.enterprise.module_6_2.model.EmployeeDAO;
+import ua.goit.timonov.enterprise.module_6_2.dao.EmployeeDAO;
 
 import java.util.ArrayList;
 import java.util.Date;
